backend/pyproject.toml
[project]
name = "retake-backend"
version = "0.1.0"
description = "Backend service for Retake AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "selenium>=4.17.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "chromadb>=0.4.22",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "webdriver-manager>=4.0.2",
    "playwright>=1.57.0",
    "supabase>=2.27.0",
    "postgrest>=2.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
